teensy 4.0 init

+++ lib/SoftTimer/src/BlinkTask.h
@@ -110,7 +110,7 @@ class BlinkTask : public Task
 
     uint8_t _bitMask;
 #ifndef ESP8266
-    volatile uint8_t *_portRegister;
+    volatile uint32_t *_portRegister;
 #else
     volatile uint32_t *_portRegister;
 #endif
