+++ lib/SoftTimer/src/SoftPwmTask.h
@@ -63,7 +63,7 @@ class SoftPwmTask : public Task
     
     uint8_t _bitMask;
 #ifndef ESP8266
-    volatile uint8_t *_portRegister;
+    volatile uint32_t *_portRegister;
 #else
     volatile uint32_t *_portRegister;
 #endif
